move map & pg_block_prob
