+++ pgdrive/envs/marl_envs/marl_parking_lot.py
@@ -2,12 +2,12 @@ import copy
 import logging
 
 from pgdrive.envs.marl_envs.marl_inout_roundabout import LidarStateObservationMARound
-from pgdrive.envs.multi_agent_pgdrive import MultiAgentPGDrive, pygame_replay, panda_replay
+from pgdrive.envs.multi_agent_pgdrive import MultiAgentPGDrive, panda_replay
 from pgdrive.obs.observation_base import ObservationBase
 from pgdrive.scene_creator.blocks.first_block import FirstBlock
 from pgdrive.scene_creator.blocks.parking_lot import ParkingLot
 from pgdrive.scene_creator.blocks.t_intersection import TInterSection
-from pgdrive.scene_creator.map import PGMap
+from pgdrive.scene_creator.map.pg_map import PGMap
 from pgdrive.scene_creator.road.road import Road
 from pgdrive.scene_manager.spawn_manager import SpawnManager
 from pgdrive.utils import get_np_random, PGConfig
