+++ pgdrive/envs/marl_envs/marl_tollgate.py
@@ -7,7 +7,7 @@ from pgdrive.obs.state_obs import LidarStateObservation, StateObservation
 from pgdrive.scene_creator.blocks.bottleneck import Merge, Split
 from pgdrive.scene_creator.blocks.first_block import FirstBlock
 from pgdrive.scene_creator.blocks.tollgate import TollGate
-from pgdrive.scene_creator.map import PGMap
+from pgdrive.scene_creator.map.pg_map import PGMap
 from pgdrive.scene_creator.road.road import Road
 from pgdrive.utils import PGConfig, clip
