+++ pgdrive/envs/pgdrive_env.py
@@ -2,7 +2,6 @@ import copy
 import json
 import logging
 import os.path as osp
-import sys
 from typing import Union, Dict, AnyStr, Optional, Tuple
 from pgdrive.utils.engine_utils import pgdrive_engine_initialized
 import numpy as np
@@ -11,7 +10,8 @@ from pgdrive.envs.base_env import BasePGDriveEnv
 from pgdrive.obs.image_obs import ImageStateObservation
 from pgdrive.obs.state_obs import LidarStateObservation
 from pgdrive.scene_creator.blocks.first_block import FirstBlock
-from pgdrive.scene_creator.map import Map, MapGenerateMethod, parse_map_config, PGMap
+from pgdrive.scene_creator.map.map import Map, MapGenerateMethod, parse_map_config
+from pgdrive.scene_creator.map.pg_map import PGMap
 from pgdrive.scene_creator.vehicle.base_vehicle import BaseVehicle
 from pgdrive.scene_creator.vehicle_module.distance_detector import DetectorMask
 from pgdrive.scene_manager.traffic_manager import TrafficMode
