+++ pgdrive/scene_creator/map/city_map.py
@@ -5,8 +5,8 @@ from panda3d.core import NodePath
 from pgdrive.scene_creator.algorithm.BIG import BIG
 from pgdrive.scene_creator.blocks.block import Block
 from pgdrive.scene_creator.blocks.first_block import FirstBlock
-from pgdrive.scene_creator.map import Map
-from pgdrive.scene_creator.pg_blocks import PGBlock
+from pgdrive.scene_creator.map.map import Map
+from pgdrive.scene_creator.algorithm.blocks_prob_dist import PGBlock
 from pgdrive.scene_creator.road.road_network import RoadNetwork
 from pgdrive.engine.world.pg_physics_world import PGPhysicsWorld
