+++ pgdrive/tests/scripts/capture_obs.py
@@ -1,5 +1,5 @@
 from pgdrive.envs.pgdrive_env import PGDriveEnv
-from pgdrive.scene_creator.map import Map, MapGenerateMethod
+from pgdrive.scene_creator.map.map import Map, MapGenerateMethod
 from pgdrive.scene_creator.vehicle_module.depth_camera import DepthCamera
 from pgdrive.utils import setup_logger
