+++ pgdrive/tests/test_funtionality/_test_record_restore_marl.py
@@ -1,9 +1,6 @@
 import json
 
 from pgdrive.envs.marl_envs.marl_inout_roundabout import MultiAgentRoundaboutEnv
-from pgdrive.envs.marl_envs.marl_intersection import MultiAgentIntersectionEnv
-from pgdrive.scene_creator.map import Map, MapGenerateMethod
-from pgdrive.scene_manager.traffic_manager import TrafficMode
 from pgdrive.utils import setup_logger
